move binders to console project and use temp file for arguments
